+++ backend/src/app.module.ts
@@ -3,9 +3,10 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { HealthModule } from './health/health.module';
 import { PrismaModule } from './prisma/prisma.module';
-
+import { AuthModule } from './auth/auth.module';
+import { AdminModule } from './admin/admin.module';
 @Module({
-  imports: [HealthModule, PrismaModule],
+  imports: [HealthModule, PrismaModule, AuthModule, AdminModule],
   controllers: [AppController],
   providers: [AppService],
 })